cleaning up some changes
