+++ quadratic-core/Cargo.toml
@@ -21,7 +21,7 @@ name = "export_types"
 path = "src/bin/export_types.rs"
 
 [features]
-default = ["console_error_panic_hook", "js"]
+default = ["console_error_panic_hook", "js", "show-operations"]
 # "js" feature is disabled for testing (particularly WASI benchmarks)
 js = ["js-sys", "serde-wasm-bindgen", "ts-rs", "wasm-bindgen", "wasm-bindgen-futures"]
 show-operations = []
