+++ quadratic-core/src/controller/execution/execute_operation.rs
@@ -348,81 +348,3 @@ impl GridController {
         }
     }
 }
-
-/*
-
-    /*
-
-
-    #[test]
-    fn test_spilled_output_over_code_cell() {
-        let mut gc = GridController::new();
-        let sheet_id = gc.sheet_ids()[0];
-
-        let code_cell_output = Some(CodeCellValue {
-            language: CodeCellLanguage::Python,
-            code_string: "print(1)".to_string(),
-            formatted_code_string: None,
-            last_modified: String::from(""),
-            output: Some(CodeCellRunOutput {
-                std_out: None,
-                std_err: None,
-                result: crate::grid::CodeCellRunResult::Ok {
-                    output_value: Value::Array(Array::new_empty(
-                        ArraySize::try_from((2, 3)).expect("failed to create array"),
-                    )),
-                    cells_accessed: HashSet::new(),
-                },
-                spill: false,
-            }),
-        });
-
-        let original = SheetPos {
-            x: 0,
-            y: 1,
-            sheet_id,
-        };
-        gc.transaction_in_progress = true;
-        gc.update_code_cell_value(original, code_cell_output.clone());
-
-        let code_cell_output = Some(CodeCellValue {
-            language: CodeCellLanguage::Python,
-            code_string: "print(1)".to_string(),
-            formatted_code_string: None,
-            last_modified: String::from(""),
-            output: Some(CodeCellRunOutput {
-                std_out: None,
-                std_err: None,
-                result: crate::grid::CodeCellRunResult::Ok {
-                    output_value: Value::Array(Array::new_empty(
-                        ArraySize::try_from((1, 3)).expect("failed to create array"),
-                    )),
-                    cells_accessed: HashSet::new(),
-                },
-                spill: false,
-            }),
-        });
-
-        let sheet_pos = SheetPos {
-            x: 0,
-            y: 0,
-            sheet_id,
-        };
-
-        gc.update_code_cell_value(sheet_pos, code_cell_output.clone());
-
-        let sheet = gc.grid.sheet_from_id(sheet_id);
-        let code_cell = sheet.get_code_cell(Pos { x: 0, y: 0 });
-        assert!(code_cell.unwrap().has_spill_error());
-        assert_eq!(
-            sheet.get_column(0).unwrap().spills.get(0),
-            Some(sheet_pos.into())
-        );
-        assert_eq!(
-            sheet.get_column(0).unwrap().spills.get(1),
-            Some(original.into())
-        );
-    }
-     */
-}
- */
