+++ quadratic-core/src/controller/execution/tests/mod.rs
@@ -1,3 +1,4 @@
+mod cell_value;
 mod offsets;
 mod sheets;
 mod spills;
